Design: 캐러셀을 버튼을 사용해 다음,이전으로 이동하도록 UI 수정
